+++ salbox-landing/src/modules/Register.tsx
@@ -1,15 +1,19 @@
+import { Dispatch } from 'react';
 import RegisterImg from "../assets/images/RegisterImg.svg";
+import RegisterForm from '../components/RegisterForm';
 
 type RegisterProps = {
   phoneNumber: string | null;
-  handlePhoneNumber: any;
+  handlePhoneNumber: () => void;
   registered: boolean;
+  setRegistered: Dispatch<React.SetStateAction<boolean>>;
 };
 
 const Register = ({
   phoneNumber,
   handlePhoneNumber,
   registered,
+  setRegistered
 }: RegisterProps) => {
   return (
     <main className="bg-moonstone-400 max-w-7xl m-4 mx-4 xl:mx-auto px-8 pt-10 lg:pt-4 text-center h-full lg:text-start rounded-xl flex flex-col lg:gap-10 items-center justify-center lg:flex-row-reverse lg:my-10 lg:pr-20">
@@ -22,53 +26,11 @@ const Register = ({
             Neque porro quisquam est qui dolorem ipsum quia dolor sit amet,
             consectetur, adipisci velit Neque porro quisquam.
           </p>
-          {/* Form for mobile */}
-          <form className="flex items-center mb-5 sm:hidden">
-            <div className="flex items-center bg-white rounded-full rounded-r-none py-3 px-1">
-              <img
-                className="w-6 h-6 ml-5"
-                src="icons/IconWhatsapp.png"
-                alt="Hero Image"
-              />
-              <input
-                className="ml-8 text-base w-[160px] focus:outline-none bg-transparent"
-                type="text"
-                placeholder="0000000000"
-                maxLength={10}
-                value={phoneNumber ? phoneNumber : ""}
-                onChange={handlePhoneNumber}
-              />
-            </div>
-            <input
-              className="bg-scarlet-400 text-lg font-semibold rounded-full rounded-l-none border-none px-6 py-3 h-12"
-              type="image"
-              src="icons/IconSend.png"
-              alt="Send Button"
-            />
-          </form>
-          {/* Form for desktop */}
-          <div className="hidden sm:flex items-center bg-white rounded-full py-1 px-1 mb-10">
-            <img
-              className="w-6 h-6 ml-5"
-              src="icons/IconWhatsapp.png"
-              alt="Hero Image"
-            />
-            <form>
-              <input
-                className="ml-8 text-xl w-[180px] focus:outline-none bg-transparent"
-                type="text"
-                placeholder="000 000 0000"
-                maxLength={10}
-                value={phoneNumber ? phoneNumber : ""}
-                onChange={handlePhoneNumber}
-              />
-              <input
-                className="bg-scarlet-400 text-white text-lg font-semibold rounded-full border-none px-16 py-2"
-                type="submit"
-                value="Enviar"
-              />
-            </form>
-          </div>
+          <RegisterForm
+            phoneNumber={phoneNumber}
+            handlePhoneNumber={handlePhoneNumber}
+            setRegistered={setRegistered}
+          />
         </section>
       ) : (
         <section className="flex flex-col items-center lg:items-start">
@@ -76,7 +38,7 @@ const Register = ({
             ¡Gracias por Registrarte con Salbox!
           </h2>
           <p className="my-4 text-sm md:text-base md:w text-white leading-8 md:leading-8 md:my-8">
-            Revisa tu whatsapp para descubir lo que Salbox tiene para ofrecerte.
+            Te redireccionaremos al WhatsApp de Salbox ahora...
           </p>
         </section>
       )}
